+++ common/src/main/java/dev/architectury/event/events/common/EntityEvent.java
@@ -22,7 +22,6 @@ package dev.architectury.event.events.common;
 import dev.architectury.event.Event;
 import dev.architectury.event.EventFactory;
 import dev.architectury.event.EventResult;
-import net.minecraft.world.InteractionResult;
 import net.minecraft.world.InteractionResultHolder;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.Entity;
@@ -57,7 +56,7 @@ public interface EntityEvent {
      */
     Event<EnterSection> ENTER_SECTION = EventFactory.createLoop();
     /**
-     * @see AnimalTame#onTame(Animal, Player)
+     * @see AnimalTame#tame(Animal, Player)
      */
     Event<AnimalTame> ANIMAL_TAME = EventFactory.createEventResult();
     
@@ -147,8 +146,8 @@ public interface EntityEvent {
          * @param animal The animal being tamed.
          * @param player The tamer.
          * @return A {@link EventResult} determining the outcome of the event,
-         * if the outcome is EventResult::interruptFalse, the animal isn't tamed.
+         * the action may be cancelled by the result.
          */
-        EventResult onTame(Animal animal, Player player);
+        EventResult tame(Animal animal, Player player);
     }
 }
