+++ fabric/src/main/java/dev/architectury/mixin/fabric/MixinOcelot.java
@@ -20,31 +20,26 @@
 package dev.architectury.mixin.fabric;
 
 import dev.architectury.event.events.common.EntityEvent;
-import net.minecraft.world.entity.ai.goal.RunAroundLikeCrazyGoal;
-import net.minecraft.world.entity.animal.horse.AbstractHorse;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.entity.animal.Animal;
+import net.minecraft.world.entity.animal.Ocelot;
 import net.minecraft.world.entity.player.Player;
-import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-@Mixin(RunAroundLikeCrazyGoal.class)
-public class HorseTameInvoker {
-    @Shadow
-    @Final
-    private AbstractHorse horse;
-    
-    @Inject(method = "tick",
+@Mixin(Ocelot.class)
+public class MixinOcelot {
+    @Inject(method = "mobInteract",
             at = @At(
                     value = "INVOKE",
-                    target = "Lnet/minecraft/world/entity/animal/horse/AbstractHorse;tameWithName(Lnet/minecraft/world/entity/player/Player;)Z"
-            ), cancellable = true
-    )
-    private void tick(CallbackInfo ci) {
-        if (EntityEvent.ANIMAL_TAME.invoker().tame(this.horse, (Player) this.horse.getPassengers().get(0)).isFalse()) {
-            ci.cancel();
+                    target = "Lnet/minecraft/world/entity/animal/Ocelot;setTrusting(Z)V"
+            ), cancellable = true)
+    private void mobInteract(Player player, InteractionHand hand, CallbackInfoReturnable<InteractionResult> cir) {
+        if (EntityEvent.ANIMAL_TAME.invoker().tame((Animal) (Object) this, player).isFalse()) {
+            cir.setReturnValue(InteractionResult.PASS);
         }
     }
 }
