+++ fabric/src/main/java/dev/architectury/mixin/fabric/MixinTamableAnimal.java
@@ -20,30 +20,19 @@
 package dev.architectury.mixin.fabric;
 
 import dev.architectury.event.events.common.EntityEvent;
-import net.minecraft.world.entity.ai.goal.RunAroundLikeCrazyGoal;
-import net.minecraft.world.entity.animal.horse.AbstractHorse;
+import net.minecraft.world.entity.TamableAnimal;
+import net.minecraft.world.entity.animal.Animal;
 import net.minecraft.world.entity.player.Player;
-import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-@Mixin(RunAroundLikeCrazyGoal.class)
-public class HorseTameInvoker {
-    @Shadow
-    @Final
-    private AbstractHorse horse;
-    
-    @Inject(method = "tick",
-            at = @At(
-                    value = "INVOKE",
-                    target = "Lnet/minecraft/world/entity/animal/horse/AbstractHorse;tameWithName(Lnet/minecraft/world/entity/player/Player;)Z"
-            ), cancellable = true
-    )
-    private void tick(CallbackInfo ci) {
-        if (EntityEvent.ANIMAL_TAME.invoker().tame(this.horse, (Player) this.horse.getPassengers().get(0)).isFalse()) {
+@Mixin(TamableAnimal.class)
+public class MixinTamableAnimal {
+    @Inject(method = "tame", at = @At(value = "HEAD"), cancellable = true)
+    private void tame(Player player, CallbackInfo ci) {
+        if (EntityEvent.ANIMAL_TAME.invoker().tame((Animal) (Object) this, player).isFalse()) {
             ci.cancel();
         }
     }
