+++ testmod-common/src/main/java/dev/architectury/test/events/DebugEvents.java
@@ -36,7 +36,6 @@ import net.minecraft.core.Vec3i;
 import net.minecraft.network.chat.TextComponent;
 import net.minecraft.network.chat.TranslatableComponent;
 import net.minecraft.world.InteractionHand;
-import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.player.Player;
@@ -119,8 +118,7 @@ public class DebugEvents {
             return EventResult.pass();
         }));
         EntityEvent.ANIMAL_TAME.register(((animal, player) -> {
-            TestMod.SINK.accept("%s tamed %s at %s", player.getScoreboardName(), animal.getDisplayName().getString(), animal.position());
-            player.displayClientMessage(new TextComponent(player.getScoreboardName() + " tamed " + animal.getDisplayName().getString() + " at " + animal.position()), true);
+            TestMod.SINK.accept("%s tamed %s at %s", player.getScoreboardName(), animal.getDisplayName().getString(), toShortString(animal.position()));
             return EventResult.pass();
         }));
         ExplosionEvent.DETONATE.register((world, explosion, affectedEntities) -> {
